+++ packages/ds/angular/src/stencil-generated/components.ts
@@ -29,6 +29,48 @@ export class DjlButton {
 }
 
 
+export declare interface DjlIcon extends Components.DjlIcon {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['iconConfig']
+})
+@Component({
+  selector: 'djl-icon',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['iconConfig']
+})
+export class DjlIcon {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
+export declare interface DjlLink extends Components.DjlLink {}
+
+@ProxyCmp({
+  defineCustomElementFn: undefined,
+  inputs: ['ariaLabel', 'download', 'href', 'target']
+})
+@Component({
+  selector: 'djl-link',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  template: '<ng-content></ng-content>',
+  inputs: ['ariaLabel', 'download', 'href', 'target']
+})
+export class DjlLink {
+  protected el: HTMLElement;
+  constructor(c: ChangeDetectorRef, r: ElementRef, protected z: NgZone) {
+    c.detach();
+    this.el = r.nativeElement;
+  }
+}
+
+
 export declare interface DjlFormCheckbox extends Components.DjlFormCheckbox {}
 
 @ProxyCmp({