+++ packages/ds/react/lib/stencil-generated/components.ts
@@ -8,5 +8,7 @@ import type { JSX } from '@ircc-ca/ds-sdc-web';
 
 
 export const DjlButton = /*@__PURE__*/createReactComponent<JSX.DjlButton, HTMLDjlButtonElement>('djl-button');
+export const DjlIcon = /*@__PURE__*/createReactComponent<JSX.DjlIcon, HTMLDjlIconElement>('djl-icon');
+export const DjlLink = /*@__PURE__*/createReactComponent<JSX.DjlLink, HTMLDjlLinkElement>('djl-link');
 export const DjlFormCheckbox = /*@__PURE__*/createReactComponent<JSX.DjlFormCheckbox, HTMLDjlFormCheckboxElement>('djl-form-checkbox');
 export const MyComponent = /*@__PURE__*/createReactComponent<JSX.MyComponent, HTMLMyComponentElement>('my-component');